+++ neoforge/src/test/java/net/minecraft/world/item/equipment/Equippable.java
@@ -1,0 +1,6 @@
+package net.minecraft.world.item.equipment;
+
+import net.minecraft.world.entity.EquipmentSlot;
+
+public record Equippable(EquipmentSlot slot) {
+}